# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BdShop",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "petstore.swagger.io",
            "description": "Api server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development-server"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Returns all users list",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Anik Ahamed"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "anik023"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "anik023@gmail.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "dkakdnahwhheohwhdhsb"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create an user",
                "description": "Create an new user",
                "operationId": "create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Binary content of file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfull"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "anik023@gmail.com"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User data",
                "description": "Update existing user data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfull operation"
                    },
                    "400": {
                        "description": "Error validation"
                    }
                }
            }
        }
    }
}